requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.14
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.1.2
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
pillow==11.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
pytz==2025.2
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
textblob==0.19.0
tiktoken==0.9.0
torch==2.7.1
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
